Cleanup test helpers
